* enc/utf_16le.c: surpress warning: shorten-64-to-32.
* ext/dbm/dbm.c: ditto.

* ext/gdbm/gdbm.c: ditto.

* parse.y (Init_ripper): surpress warning: unused value.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@31141 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ enc/utf_16le.c
@@ -56,7 +56,7 @@ static int
 utf16le_mbc_enc_len(const UChar* p, const OnigUChar* e,
 		    OnigEncoding enc ARG_UNUSED)
 {
-  ptrdiff_t len = e - p;
+  int len = (int)(e - p);
   UChar byte;
   if (len < 2)
     return ONIGENC_CONSTRUCT_MBCLEN_NEEDMORE(1);
